requirements.txt
dacite==1.8.0
MLB_StatsAPI==1.6.1
numpy==1.24.2
pandas==1.5.3
pyowm==3.3.0
SQLAlchemy==2.0.7
tqdm==4.64.0
